backend/requirements.txt
fastapi
pydantic
uvicorn
mysql-connector-python